+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added Version filter parameter for `Get-LocalInstance`.
+- Parameter to create a SqlClientConnection in `New-LocalInstance`.
 
 ## [1.5.0] - 2022-09-30